# swagger/swagger.yaml
swagger: '2.0'
info:
  description: Aplicación dirigida a compañías de Marketing para administrar todos los contactos de sus clientes para sus campañas.
  version: 1.0.0
  title: Data Warehouse
  contact:
    email: solmartinezpisani@gmail.com
...
schemes:
  - https
  - http
paths:
...
  /companias:
    get: 
      tags:
        - companias
      summary: Ver compañías
      operationId: getCompanias
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetCompanias'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    post:
      tags:
        - companias
      summary: Crear compañía
      operationId: createCompania
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Crear compañía
          required: true
          schema:
            $ref: '#/definitions/CreateCompanias'
      responses:
        '200':
          description: Successful Register
          schema:
            type: object
            properties:
              message:
                type: string
                example: Compañía agregada
        '400':
          description: Unsuccessful Register
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 400
              message:
                type: string
                example: Compañía ya guardada
  /companias/{id}:
    get:
      tags:
        - companias
      summary: Ver una sola compañía
      operationId: getCompania
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the company.
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetCompanias'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    patch:
      tags:
        - companias
      summary: Editar compañía
      operationId: updateComp
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the company to update.
        - in: body
          name: body
          description: >-
            Editar compañía
          required: true
          schema:
            $ref: '#/definitions/CreateCompanias'
      responses:
        '200':
          description: Successful operation
          schema:
            properties:
              message:
                type: string
                example: Compañia actualizada
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    delete: 
      tags:
        - companias
      summary: Eliminar compañía
      operationId: deleteComp
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the company to delete.
      responses:
        '200':
          description: Successful operation
          schema:
            type: object
            properties:
              message:
                type: string
                example: Compañía eliminada
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404                
...
definitions:
  Authenticate:
    type: object
    properties:
      message:
        type: string
        example: Header must contain JWT to verify identity.
...
  GetCompanias:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      nombre:
        type: string
        example: Acamica
      pais:
        type: string
        example: Argentina
      email:
        type: string
        example: hola@acamica.com
      direccion: 
        type: string
        example: Av. Corrientes 1233
      telefono:
        type: integer
        format: int64
        example: 45568899
...
  CreateCompanias:
    type: object
    properties:
      nombre:
        type: string
        example: Acamica
      pais:
        type: string
        example: Argentina
      email:
        type: string
        example: hola@acamica.com
      direccion: 
        type: string
        example: Av. Corrientes 1233
      telefono:
        type: integer
        format: int64
        example: 45568899
...